+++ core/modules/url/store/actions.ts
@@ -18,6 +18,8 @@ import * as cmsPageMutationTypes from '@vue-storefront/core/modules/cms/store/pa
 import isEqual from 'lodash-es/isEqual'
 import * as types from './mutation-types'
 import omit from 'lodash-es/omit'
+import { storeProductToCache } from '@vue-storefront/core/modules/catalog/helpers/search';
+import { prepareProducts } from '@vue-storefront/core/modules/catalog/helpers/prepare';
 
 // it's a good practice for all actions to return Promises with effect of their execution
 export const actions: ActionTree<UrlState, any> = {
@@ -74,7 +76,7 @@ export const actions: ActionTree<UrlState, any> = {
     })
   },
   /**
-   * @deprecated from 1.13
+   * @deprecated from 1.12
    * Router mapping fallback - get the proper URL from API
    * This method could be overriden in custom module to provide custom URL mapping logic
    */
@@ -83,31 +85,17 @@ export const actions: ActionTree<UrlState, any> = {
       Deprecated action mappingFallback - use mapFallbackUrl instead.
       You can enable mapFallbackUrl by changing 'config.urlModule.enableMapFallbackUrl' to true
     `)()
-    const { storeCode, appendStoreCode } = currentStoreView()
     const productQuery = new SearchQuery()
     url = (removeStoreCodeFromRoute(url.startsWith('/') ? url.slice(1) : url) as string)
     productQuery.applyFilter({ key: 'url_path', value: { 'eq': url } }) // Tees category
     const products = await dispatch('product/list', { query: productQuery }, { root: true })
     if (products && products.items && products.items.length) {
-      Logger.log(url, 'search', productQuery)()
       const product = products.items[0]
-      return {
-        name: localizedDispatcherRouteName(product.type_id + '-product', storeCode, appendStoreCode),
-        params: {
-          slug: product.slug,
-          parentSku: product.sku,
-          childSku: params['childSku'] ? params['childSku'] : product.sku
-        }
-      }
+      return transformProductUrl(product, params)
     } else {
       const category = await dispatch('category/single', { key: 'url_path', value: url }, { root: true })
       if (category !== null) {
-        return {
-          name: localizedDispatcherRouteName('category', storeCode, appendStoreCode),
-          params: {
-            slug: category.slug
-          }
-        }
+        return transformCategoryUrl(category)
       }
     }
   },
@@ -119,7 +107,7 @@ export const actions: ActionTree<UrlState, any> = {
     url = (removeStoreCodeFromRoute(url.startsWith('/') ? url.slice(1) : url) as string)
 
     // search for record in ES based on `url`
-    const fallbackData = await dispatch('getFallbackByUrl', { url })
+    const fallbackData = await dispatch('getFallbackByUrl', { url, params })
 
     // if there is record in ES then map data
     if (fallbackData) {
@@ -140,7 +128,9 @@ export const actions: ActionTree<UrlState, any> = {
   /**
    * Search for record in ES which contains url value (check which fields it searches in vsf-api config.urlModule.map.searchedFields)
    */
-  async getFallbackByUrl (context, { url }) {
+  async getFallbackByUrl (context, { url, params }) {
+    const groupId = (config.usePriceTiers && context.rootState.user.groupId) || null
+    const groupToken = context.rootState.user.groupToken || null
     try {
       const requestUrl = `${adjustMultistoreApiUrl(processURLAddress(config.urlModule.map_endpoint))}`
       let response: any = await fetch(
@@ -155,7 +145,20 @@ export const actions: ActionTree<UrlState, any> = {
           body: JSON.stringify({
             url,
             includeFields: null, // send `includeFields: null || undefined` to fetch all fields
-            excludeFields: []
+            excludeFields: [],
+            options: {
+              prefetchGroupProducts: true,
+              assignProductConfiguration: true,
+              populateRequestCacheTags: false,
+              setProductErrors: false,
+              fallbackToDefaultWhenNoAvailable: true,
+              separateSelectedVariant: false,
+              setConfigurableProductOptions: config.cart.setConfigurableProductOptions,
+              filterUnavailableVariants: config.products.filterUnavailableVariants
+            },
+            filters: { sku: params.childSku },
+            groupId,
+            groupToken
           })
         }
       )
@@ -199,7 +202,8 @@ export const actions: ActionTree<UrlState, any> = {
   async saveFallbackData ({ commit }, { _type, _source }) {
     switch (_type) {
       case 'product': {
-        // TODO: find a way to cache simple, configurable, group, bundle or custom_options products
+        const [product] = prepareProducts([_source])
+        storeProductToCache(product, 'sku')
         break
       }
       case 'category': {