+++ src/modules/icmaa-monitoring/server.ts
@@ -1,11 +1,26 @@
 import { serverHooks } from '@vue-storefront/core/server/hooks'
-import { icmaa_monitoring } from 'config'
+import config, { icmaa_monitoring } from 'config'
 import createLogger, { createExceptionLogger } from './lib/winston/logger'
 
-// Enable tracing by datadog
-import './lib/datadog/dd-trace'
-
 if (icmaa_monitoring.datadog.enabled === true) {
+  // Enable datadog tracer
+  const tracer = require('dd-trace')
+  tracer.init({
+    service: 'vue-storefront',
+    env: config.icmaa.mandant + '-' + config.icmaa.environment,
+    clientToken: process.env.DD_CLIENT_TOKEN || config.icmaa_monitoring.datadog.clientToken,
+    analytics: true,
+    logInjection: true,
+    runtimeMetrics: true,
+    reportHostname: true,
+    tags: {
+      env: config.icmaa.environment,
+      mandant: config.icmaa.mandant,
+      application_name: 'vue-storefront',
+      vsf: true
+    }
+  })
+
   // Enable exception-logging for unhandled ssr exceptions
   serverHooks.afterProcessStarted(() => {
     createExceptionLogger()
@@ -13,7 +28,12 @@ if (icmaa_monitoring.datadog.enabled === true) {
 
   // Log handled ssr exeptions using winston
   serverHooks.ssrException(({ err }) => {
+    let payload = { payload: err as any }
+    if (err instanceof Error) {
+      payload = { payload: err.message + '\n' + err.stack }
+    }
+
     const logger = createLogger()
-    logger.log('error', 'SSR Exception', err)
+    logger.log('error', 'SSR Exception', payload)
   })
 }