+++ src/modules/icmaa-url/store/actions.ts
@@ -67,7 +67,7 @@ const forCmsPageUrls = async ({ dispatch }, { urlPath }: UrlMapperOptions) => {
  * This is our cms landing page url fallback mapper
  */
 const forCmsLandingPageUrls = async ({ dispatch }, { urlPath }: UrlMapperOptions) => {
-  return dispatch('icmaaCmsLangingPages/single', { value: urlPath }, { root: true })
+  return dispatch('icmaaCmsLandingPages/single', { value: urlPath }, { root: true })
     .then((page: GenericStateItem) => {
       if (page !== null && Object.values(page).length > 0) {
         return {
@@ -104,7 +104,7 @@ const forCmsCompetitionsUrls = async ({ dispatch }, { urlPath }: UrlMapperOption
 }
 
 export const actions: ActionTree<UrlState, any> = {
-  async mapFallbackUrl ({ dispatch }, { url, params }: { url: string, params: any}) {
+  async mapFallbackUrl ({ dispatch }, { url, params }: { url: string, params: any }) {
     const urlPath = getUrlPathFromUrl(url)
     const paramsObj = { urlPath, params }
 
@@ -114,7 +114,7 @@ export const actions: ActionTree<UrlState, any> = {
     }
 
     // This is the code of VSF
-    const fallbackData = await dispatch('getFallbackByUrl', { url: urlPath })
+    const fallbackData = await dispatch('getFallbackByUrl', { url: urlPath, params })
     if (fallbackData) {
       const [result] = await Promise.all([
         dispatch('transformFallback', { ...fallbackData, params }),