+++ src/themes/icmaa-imp/components/core/blocks/AddToCartSidebar/AddToCartSidebar.vue
@@ -39,6 +39,7 @@
 import i18n from '@vue-storefront/i18n'
 import { mapGetters } from 'vuex'
 import { notifications } from '@vue-storefront/core/modules/cart/helpers'
+import { filterChangedProduct } from '@vue-storefront/core/modules/catalog/events'
 import Composite from '@vue-storefront/core/mixins/composite'
 import ProductPriceMixin from 'theme/mixins/product/priceMixin'
 import ProductOptionsMixin from 'theme/mixins/product/optionsMixin'
@@ -73,7 +74,6 @@ export default {
   computed: {
     ...mapGetters({
       product: 'product/getCurrentProduct',
-      originalProduct: 'product/getOriginalProduct',
       configuration: 'product/getCurrentProductConfiguration',
       options: 'product/getCurrentProductOptions',
       isAddingToCart: 'cart/getIsAdding'
@@ -108,11 +108,18 @@ export default {
     },
     async onSelect (option) {
       if (option.available) {
-        await this.$store.dispatch('product/updateConfiguration', { option })
+        this.loading = true
 
+        // We need to set the new configuration here already to enable the loading state for the selected option
+        this.$store.dispatch('product/updateConfiguration', { option })
         this.setSelectedOptionByCurrentConfiguration()
+
+        const configuration = Object.assign({ attribute_code: option.type }, option)
+        await filterChangedProduct(configuration, this.$store, this.$router)
+
         this.$bus.$emit('user-has-selected-product-variant')
 
+        this.loading = false
         this.$store.dispatch('ui/closeAll')
       } else {
         this.selectedOption = option
@@ -126,7 +133,7 @@ export default {
       if (this.hasMultiplePrices) {
         const product = this.product.configurable_children.find(child => child[option.type] === option.id)
         if (product) {
-          return product.original_price_incl_tax
+          return product.price_incl_tax || product.original_price_incl_tax
         }
       }